feat: implement WebShell with theme toggle and smart authentication
- Replace (app)/page.tsx with redirect to prio dashboard
- Add new (site)/page.tsx with WebShell, theme toggle, and smart auth-aware button
- Disable authentication checks temporarily for development
- Enable left sidebar with expanded navigation in prio layout
- Add logo, breadcrumb, theme toggle, and user button to topbar
- Update landing page copy to emphasize collaborative AI approach
- Fix GenericBar section mapping and chat input space key handling in submodule
